Fixed breaking changes with Function references (v2)
