--- src/Apprentice.BotV4/State/UserInfo.cs
@@ -1,12 +1,0 @@
-﻿namespace ESFA.DAS.ProvideFeedback.Apprentice.BotV4.State
-{
-    using ESFA.DAS.ProvideFeedback.Apprentice.BotV4.Models;
-
-    /// <summary>
-    /// User state information.
-    /// </summary>
-    public class UserInfo
-    {
-        public SurveyState SurveyState { get; set; }
-    }
-}
