+++ src/Apprentice.Functions.NotifyMessageHandlerV2/ESFA.DAS.ProvideFeedback.Apprentice.Functions.NotifyMessageHandlerV2.csproj
@@ -4,9 +4,15 @@
     <AzureFunctionsVersion>v2</AzureFunctionsVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="1.0.13" />
+    <!--<PackageReference Include="Microsoft.Azure.WebJobs" Version="3.0.0-beta8" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Core" Version="3.0.0-beta8" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Storage" Version="3.0.0-beta8" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="3.0.0-beta8" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.ServiceBus" Version="3.0.0-beta8" />
+    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="1.0.19" />-->
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.7.2" />
     <PackageReference Include="Notify" Version="2.1.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
@@ -14,12 +20,11 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Clients\" />
-    <Folder Include="Properties\PublishProfiles\" />
-  </ItemGroup>
-  <ItemGroup>
+    <ProjectReference Include="..\Apprentice.Functions\ESFA.DAS.ProvideFeedback.Apprentice.Functions.csproj" />
     <ProjectReference Include="..\Apprentice.Core\ESFA.DAS.ProvideFeedback.Apprentice.Core.csproj" />
     <ProjectReference Include="..\Apprentice.Data\ESFA.DAS.ProvideFeedback.Apprentice.Data.csproj" />
-    <ProjectReference Include="..\Apprentice.Functions\ESFA.DAS.ProvideFeedback.Apprentice.Functions.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Properties\" />
   </ItemGroup>
 </Project>
