+++ src/Apprentice.Services.NotifySmsService/ESFA.DAS.ProvideFeedback.Apprentice.Services.NotifySmsService.csproj
@@ -5,13 +5,14 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.CSharp" Version="4.5.0" />
     <PackageReference Include="Notify" Version="2.1.0" />
     <PackageReference Include="SFA.DAS.Notifications.Api.Client" Version="2.1.12" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="2.1.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Apprentice.Core\ESFA.DAS.ProvideFeedback.Apprentice.Core.csproj" />
-    <ProjectReference Include="..\Apprentice.Infrastructure\ESFA.DAS.ProvideFeedback.Apprentice.Infrastructure.csproj" />
     <ProjectReference Include="..\Apprentice.Services\ESFA.DAS.ProvideFeedback.Apprentice.Services.csproj" />
   </ItemGroup>
